Move `backend.ObjectInfo` into `zbstore` package
Updates #345
