+++ zbstore/object_info_test.go
@@ -1,15 +1,13 @@
 // Copyright 2025 The zb Authors
 // SPDX-License-Identifier: MIT
 
-package backend_test
+package zbstore
 
 import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	. "zb.256lights.llc/pkg/internal/backend"
 	"zb.256lights.llc/pkg/sets"
-	"zb.256lights.llc/pkg/zbstore"
 	"zombiezen.com/go/nix"
 )
 
@@ -28,10 +26,10 @@ func objectInfoMarshalTests(tb testing.TB) []*objectInfoMarshalTest {
 				"NarSize: 1228440\n" +
 				"CA: fixed:r:sha256:143sdn30fdykpz8gpyw45m9m6m4gz858w9kc6myy7p0v74v5qq4m\n",
 			info: &ObjectInfo{
-				StorePath: "/zb/store/z5yrbqk8sjlzyvw8wpicsn2ybk0sc470-busybox-1.36.1",
-				NARHash:   mustParseHash(tb, "sha256:1d99d4f5hjl24w30hwgrmn00kryvd1yxvyydpkm76hgmcig9mllc"),
-				NARSize:   1228440,
-				CA:        nix.RecursiveFileContentAddress(mustParseHash(tb, "sha256:143sdn30fdykpz8gpyw45m9m6m4gz858w9kc6myy7p0v74v5qq4m")),
+				StorePath:      "/zb/store/z5yrbqk8sjlzyvw8wpicsn2ybk0sc470-busybox-1.36.1",
+				NARHash:        mustParseHash(tb, "sha256:1d99d4f5hjl24w30hwgrmn00kryvd1yxvyydpkm76hgmcig9mllc"),
+				NARSize:        1228440,
+				ContentAddress: nix.RecursiveFileContentAddress(mustParseHash(tb, "sha256:143sdn30fdykpz8gpyw45m9m6m4gz858w9kc6myy7p0v74v5qq4m")),
 			},
 		},
 		{
@@ -45,10 +43,10 @@ func objectInfoMarshalTests(tb testing.TB) []*objectInfoMarshalTest {
 				StorePath: "/zb/store/9n2ccy3mcsb04q47npp28jwkd9py3wdj-gcc-4.2.1",
 				NARHash:   mustParseHash(tb, "sha256:0a9pvsidbxbdcrj9aj3gz7sp0ibfzlhmp6jwljjqya4xjwc0lnzr"),
 				NARSize:   95499656,
-				References: *sets.NewSorted[zbstore.Path](
+				References: *sets.NewSorted[Path](
 					"/zb/store/9n2ccy3mcsb04q47npp28jwkd9py3wdj-gcc-4.2.1",
 				),
-				CA: nix.RecursiveFileContentAddress(mustParseHash(tb, "sha256:0pikk3c4s55bhh1nzbj2150ksbnsxnvl4xarap58kjp9qv5d4z1k")),
+				ContentAddress: nix.RecursiveFileContentAddress(mustParseHash(tb, "sha256:0pikk3c4s55bhh1nzbj2150ksbnsxnvl4xarap58kjp9qv5d4z1k")),
 			},
 		},
 		{
@@ -62,11 +60,11 @@ func objectInfoMarshalTests(tb testing.TB) []*objectInfoMarshalTest {
 				StorePath: "/zb/store/h9z0bp4cnw57ngi33l4781n4rrfb2a8q-gcc-4.2.1",
 				NARHash:   mustParseHash(tb, "sha256:1ksdfhnnlywdh1d0vlxrc6kdyv0vqqagkrhx2l145s5mhk2cm70z"),
 				NARSize:   16832,
-				References: *sets.NewSorted[zbstore.Path](
+				References: *sets.NewSorted[Path](
 					"/zb/store/9n2ccy3mcsb04q47npp28jwkd9py3wdj-gcc-4.2.1",
 					"/zb/store/z5yrbqk8sjlzyvw8wpicsn2ybk0sc470-busybox-1.36.1",
 				),
-				CA: nix.RecursiveFileContentAddress(mustParseHash(tb, "sha256:073lrg7m3rrqbn9wgy7wrf94h77hhhjmnvwhh8vqpnbflsgzb8dk")),
+				ContentAddress: nix.RecursiveFileContentAddress(mustParseHash(tb, "sha256:073lrg7m3rrqbn9wgy7wrf94h77hhhjmnvwhh8vqpnbflsgzb8dk")),
 			},
 		},
 		{
@@ -109,7 +107,7 @@ func TestObjectInfoUnmarshal(t *testing.T) {
 			}
 			diff := cmp.Diff(
 				test.info, got,
-				transformSortedSet[zbstore.Path](),
+				transformSortedSet[Path](),
 			)
 			if diff != "" {
 				t.Errorf("-want +got:\n%s", diff)
@@ -138,7 +136,7 @@ func FuzzObjectInfoMarshal(f *testing.F) {
 		}
 		diff := cmp.Diff(
 			got, got2,
-			transformSortedSet[zbstore.Path](),
+			transformSortedSet[Path](),
 		)
 		if diff != "" {
 			t.Errorf("round-trip (-first +second):\n%s", diff)
